refactor(post) PostsManagerPage 위치 수정
